requirements.txt
openai-whisper
sounddevice
numpy
pynput
scipy 